// apps/main/src/models/v2/controlplane/project.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { ProjectStatus } from './projectStatus';

/**
 * A project.
 */
export interface Project {
  /** Project identifier */
  id: string;
  /** Unique uid for the project to identify a unique lifecycle */
  uuid: string;
  /** Project human readable name */
  display_name: string;
  /** The date and time when the project was created */
  created_at: string;
  /** The date and time when the project was updated due to internal or external changes */
  updated_at: string;
  /** The status of the project. Projects are normally active unless they have been deleted. A deleting project is waiting for it's environments to be deleted. */
  status: ProjectStatus;
}